refactoring events to match context api
